requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Templating
jinja2==3.1.4

# Auth
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.2.0  # Starlette sessions

# Database (async)
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0         # SQLite (local dev)
asyncpg==0.29.0           # PostgreSQL (production)
alembic==1.13.1

# Storage
google-cloud-storage==2.16.0

# HWPX / ZIP manipulation
lxml==5.2.1

# Utilities
python-multipart==0.0.9   # file upload
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
